requirements.txt
# Core dependencies
langchain==0.1.0
langchain-openai==0.0.2
openai==1.12.0
python-dotenv==1.0.0

# MCP and LangGraph
mcp==0.1.0
langgraph==0.0.10
langchain-mcp-adapters==0.1.0

# FastAPI and Streamlit
fastapi==0.109.2
uvicorn==0.27.1
streamlit==1.31.1
requests==2.31.0
pydantic==2.6.1

# AWS and Pinecone
pinecone-client==3.0.2
boto3==1.34.29

# Tracing and monitoring
langsmith==0.0.87